switched to client component for main page
